refactor: remove desktop-related file into Desktop folder
