+++ packages/ui/Desktop/index.tsx
@@ -1,0 +1,6 @@
+export { DesktopHeader } from "./Header";
+export { DesktopFooter } from "./Footer";
+export { Menu } from "./Menu";
+export { SelectRect } from "./SelectRect";
+export { DesktopContainer } from "./DesktopContainer";
+export { DesktopApp } from "./App";
